@@ -1,11 +1,22 @@
 module github.com/pubgo/xerror
 
-go 1.14
+go 1.18
 
 require (
+	github.com/golang/protobuf v1.5.2
 	github.com/google/go-cmp v0.5.6
 	github.com/kr/pretty v0.3.0
 	github.com/valyala/bytebufferpool v1.0.0
 	go.uber.org/atomic v1.7.0
 	go.uber.org/goleak v1.1.10
+	google.golang.org/grpc v1.46.2
+)
+
+require (
+	github.com/kr/text v0.2.0 // indirect
+	github.com/rogpeppe/go-internal v1.6.1 // indirect
+	golang.org/x/lint v0.0.0-20190930215403-16217165b5de // indirect
+	golang.org/x/tools v0.0.0-20191108193012-7d206e10da11 // indirect
+	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
+	google.golang.org/protobuf v1.27.1 // indirect
 )
